(GH-1405) Convert resolve_reference task to use ruby_plugin_helper
This updates the `resolve_reference` task to use the ruby plugin helper
to perform operations common to Bolt plugins, such as mapping resource
values to target data.

It replaces several methods that were part of the task with those in the
plugin helper.

This also replaces the `statefile` parameter for the `resolve_reference` task
with a `state` parameter to maintain consistency among the tasks and
plans in the module.