(GH-1436) Raise error when remote state cannot be loaded
This fixes a bug that raises an obscure error caused by an attempt to
load a non-existent state file. Previously, if remote state could not be
loaded, `terraform state pull` would return successfully with a `nil`
value. This value would be loaded into the inventory, causing Bolt to
raise an error. Now, the plugin will check that `terraform state pull`
not only returned successfully but that it also returned data to act on.