Handle case of absent statefile
The plugin should not crash when the statefile is absent. Rather, it
should return no inventory.

Use case:

I am the purveyor of a Bolt project which:
 * Uses the Terraform module to provision new infrastructure
 * Uses an inventory.yaml file, and the Terraform inventory plugin, to make
   post-provisioning interaction with that infrastructure easy

When I have cloned a new, fresh copy of my project:
 * I want to be able to run bolt inventory show showing no systems before I
   have provisioned them
 * I want to be able to run bolt inventory show showing new systems, after
   provisioning
 * I want to be able to do this without modifying the inventory.yaml file I
   ship in my project

In the current design, in which the inventory plugin raises an exception
if no Terraform statefile is found, this is impossible.

After this commit, rather than raising an exception, in the event a
Terraform statefile does not exist, we interpret this as "there is no
state". In terms of inventory data, this means "there is no inventory".
This is not considered an exception, but rather a simple empty inventory
situation.