(maint) Add PDK as a dependency
Running the module release pipeline is causing the error `Errno::ENOENT:
No such file or directory - pdk`, which is the same issue described [in
this
commit](puppetlabs/puppetlabs-azure_inventory#11).
Adding PDK as a depedency should ensure that the new release rake task
succeeds.

This also preps to release the update as 0.6.1