Add refresh_state option to apply plan
Provides a refresh_state option to the terraform::apply plan that will
optionally run the terrafrom::refresh task after doing an apply. This
allows for situations where resources were modified during the apply and
not written to state so a refresh is required to capture them.

Use case example. Using the aws_ec2_tag to tag instances will not write
to state an instance resource with the tag since it is added after the
initial state sync. Running a refresh post apply will discover the tag
and write it to the instance's state entry.

Example code:
https://github.com/puppetlabs/terraform-aws-pe_arch/commit/572f1feff35af1d9cece4e7f2e787000d4be31d7